+++ src/o365/spo/commands/hubsite/QueryListResult.ts
@@ -1,0 +1,11 @@
+export interface QueryListResult {
+  FilterLink: string;
+  FirstRow: number;
+  FolderPermissions: string;
+  ForceNoHierarchy: string;
+  HierarchyHasIndention: string | null;
+  LastRow: number;
+  Row: any[];
+  NextHref: string | null;
+  RowLimit: number;
+}